feat(screen): add inventory overlay to cartridge screen
Added an overlay feature to the Cartridge Inventory Screen that visually indicates the current selected inventory. This includes enhancements to item rendering and the integration of a new utility function for drawing textures with glint effects.

Includes the addition of `DrawContextUtils` for improved texture handling and adjustment of inventory components to support the new overlay functionality.