feat(dice): add item reference to DiceEntity class
Added functionality to store and manage a reference to the item associated with the DiceEntity. This includes methods to get and set the item reference and updates to NBT serialization for saving and loading the item reference.

BREAKING CHANGE: Updated the configuration method for DiceEntity to include item reference handling.