feat(keybinds): add recipe export functionality and improved hop switch and switchy blocks
Added functionality to export crafting recipes via a keybind. Users can now press a designated key while in a crafting screen to copy the recipe in JSON format to the clipboard. This feature simplifies the process of sharing and documenting recipes.

Also updated switches and textures for improved user experience.